app/requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
requests==2.27.1
django-crispy-forms==1.14.0
pendulum==2.1.2
pdfkit==1.0.0
Jinja2==3.0.3
qrcode==7.3.1
Pillow==9.0.0
aws-requests-auth==0.4.3